services/web_services/cloud_functions/tesla_api/requirements.txt
requests==2.31.0
google-cloud-secret-manager==2.16.1
